// src/models/PodcastEpisode.ts
export type PodcastInfo = {
  wrapperType: string;
  kind: string;
  artistId: number;
  collectionId: number;
  trackId: number;
  artistName: string;
  collectionName: string;
  trackName: string;
  collectionCensoredName: string;
  trackCensoredName: string;
  artistViewUrl: string;
  collectionViewUrl: string;
  feedUrl: string;
  trackViewUrl: string;
  artworkUrl30: string;
  artworkUrl60: string;
  artworkUrl100: string;
  collectionPrice: number;
  trackPrice: number;
  collectionHdPrice: number;
  releaseDate: string;
  collectionExplicitness: string;
  trackExplicitness: string;
  trackCount: number;
  trackTimeMillis: number;
  country: string;
  currency: string;
  primaryGenreName: string;
  artworkUrl600: string;
  genreIds: string[];
  genres: string[];
};

export type PodcastEpisode = {
  country: string;
  episodeUrl: string;
  previewUrl: string;
  feedUrl: string;
  closedCaptioning: string;
  artistIds: number[];
  collectionViewUrl: string;
  genres: { name: string; id: string }[];
  episodeGuid: string;
  description: string;
  shortDescription: string;
  releaseDate: string;
  trackTimeMillis: number;
  trackId: number;
  trackName: string;
  collectionId: number;
  collectionName: string;
  artworkUrl600: string;
  artworkUrl60: string;
  artistViewUrl: string;
  contentAdvisoryRating: string;
  trackViewUrl: string;
  artworkUrl160: string;
  episodeFileExtension: string;
  episodeContentType: string;
  kind: string;
  wrapperType: string;
};

export type PodcastArtistAndEpisodesResults = {
  resultCount: number;
  results: [PodcastInfo, ...PodcastEpisode[]];
};